Activitating Services feature

+++ app/Http/Controllers/JobsController.php
@@ -321,29 +321,31 @@ class JobsController extends Controller
                 'status'=>$status
             ]);
         }
-        partsorder::where('jobno',$jobno)->delete();
+        if($request->partname!==null){
+            partsorder::where('jobno',$jobno)->delete();
 
-        for ($i=0; $i < count($request->partname); $i++) {
-            if($request->pnid[$i]==""){
-                $part = parts::create([
-                    'part_name'=>$request->partname[$i],
-                    'selling_price'=>$request->amount[$i]/$request->quantity[$i],
+            for ($i=0; $i < count($request->partname); $i++) {
+                if($request->pnid[$i]==""){
+                    $part = parts::create([
+                        'part_name'=>$request->partname[$i],
+                        'selling_price'=>$request->amount[$i]/$request->quantity[$i],
+                    ]);
+                    $partid = $part->id;
+
+                    stock::create(['part_id'=>$partid,'quantity_in_stock'=>0]);
+                }else{
+                    $partid = $request->pnid[$i];
+                }
+                $sale = partsorder::updateOrCreate(['jobno'=>$jobno, 'partsname'=>$request->partname[$i]],[
+                'customerid'=>$request->customerid,
+                'jobno'=>$jobno,
+                'partsname'=>$request->partname[$i],
+                'quantity'=>$request->quantity[$i],
+                'pdate'=>date('Y-m-d'),
+                'amount'=>$request->amount[$i],
+                'pid'=>$partid
                 ]);
-                $partid = $part->id;
-
-                stock::create(['part_id'=>$partid,'quantity_in_stock'=>0]);
-            }else{
-                $partid = $request->pnid[$i];
             }
-            $sale = partsorder::updateOrCreate(['jobno'=>$jobno, 'partsname'=>$request->partname[$i]],[
-            'customerid'=>$request->customerid,
-            'jobno'=>$jobno,
-            'partsname'=>$request->partname[$i],
-            'quantity'=>$request->quantity[$i],
-            'pdate'=>date('Y-m-d'),
-            'amount'=>$request->amount[$i],
-            'pid'=>$partid
-            ]);
         }
 
         $diag = diagnosis::updateOrCreate(['jobno'=>$jobno],$request->only(
